requirements.txt
streamlit>=1.28.0
streamlit-authenticator>=0.2.3
langgraph>=0.0.40
litellm>=1.35.0
google-cloud-firestore>=2.11.1
firebase-admin>=6.2.0
langsmith>=0.1.0
pydantic>=2.5.0
python-dotenv>=1.0.0